🐛 fix: portfolioId 동적 아이디로 경로 수정
