+++ src/app/(route)/(portfolio)/portfolio/[portfolioId]/write/_components/items/ExperienceItem.tsx
@@ -1,56 +1,49 @@
 import Image from 'next/image';
-import React from 'react';
 import Input from '../Input';
-import MotionWrapper from '../../../../../../_components/MotionWrapper';
+import MotionWrapper from '@/app/_components/MotionWrapper';
 
-interface ActivityItemProps {
+interface ExperienceItemProps {
   id: number;
   onDelete: (id: number) => void;
 }
 
-function ActivityItem({ id, onDelete }: ActivityItemProps) {
+function ExperienceItem({ id, onDelete }: ExperienceItemProps) {
   return (
     <MotionWrapper>
       <section className="relative w-full py-[20px] px-[30px] bg-gray-4 rounded-[10px] mb-4">
         <div className="grid grid-flow-col justify-stretch gap-[20px]">
           <Input
             element="input"
-            label="이름"
+            label="회사명"
             size="lg"
             type="text"
             color="white"
             maxLength={50}
-            placeholder="활동 및 교육명을 입력해주세요"
+            placeholder="회사명을 입력해주세요"
+            required
           />
           <Input
             element="input"
-            label="교육기관"
+            label="직책"
             size="lg"
             type="text"
             color="white"
             maxLength={50}
-            placeholder="교육기관명을 입력해주세요"
+            placeholder="직책을 입력해주세요"
+            required
           />
         </div>
-        <Input
-          element="input"
-          label="시작 일자"
-          size="lg"
-          type="date"
-          color="white"
-          onClick={(event: React.MouseEvent<HTMLInputElement>) => console.log('Input clicked', event)}
-        />
+        <Input element="input" label="시작 일자" size="lg" type="date" color="white" />
         <Input element="input" label="종료 일자" size="lg" type="date" color="white" />
         <Input
           element="textarea"
-          label="활동 상세 내용"
+          label="주요 업무/성과"
           size="lg"
           type="text"
           color="white"
           maxLength={2000}
-          placeholder="활동 상세 내용을 입력해주세요"
+          placeholder="주요 업무/성과를 입력해주세요"
         />
-
         <button type="button" onClick={() => onDelete(id)} className="absolute top-[20px] right-10">
           <Image src="/assets/ic-delete.svg" alt="삭제" width={25} height={25} />
         </button>
@@ -59,4 +52,4 @@ function ActivityItem({ id, onDelete }: ActivityItemProps) {
   );
 }
 
-export default ActivityItem;
+export default ExperienceItem;
