+++ src/app/(route)/(portfolio)/portfolio/[portfolioId]/write/_components/section/ActivitiesSection.tsx
